Introduce RelationTraversal and GraphTraversal types
WIP

fix iterator lost argument
