+++ traversal/iterator/GraphIterator.java
@@ -24,7 +24,7 @@ import com.vaticle.typedb.core.common.iterator.FunctionalIterator;
 import com.vaticle.typedb.core.graph.GraphManager;
 import com.vaticle.typedb.core.graph.vertex.ThingVertex;
 import com.vaticle.typedb.core.graph.vertex.Vertex;
-import com.vaticle.typedb.core.traversal.Traversal;
+import com.vaticle.typedb.core.traversal.GraphTraversal;
 import com.vaticle.typedb.core.traversal.common.Identifier;
 import com.vaticle.typedb.core.traversal.common.Identifier.Variable.Retrievable;
 import com.vaticle.typedb.core.traversal.common.VertexMap;
@@ -51,7 +51,7 @@ public class GraphIterator extends AbstractFunctionalIterator<VertexMap> {
 
     private final GraphManager graphMgr;
     private final GraphProcedure procedure;
-    private final Traversal.Parameters params;
+    private final GraphTraversal.Parameters params;
     private final Set<Retrievable> filter;
     private final Map<Identifier, FunctionalIterator<? extends Vertex<?, ?>>> iterators;
     private final Map<Identifier, Vertex<?, ?>> answer;
@@ -64,7 +64,7 @@ public class GraphIterator extends AbstractFunctionalIterator<VertexMap> {
     enum State {INIT, EMPTY, FETCHED, COMPLETED}
 
     public GraphIterator(GraphManager graphMgr, Vertex<?, ?> start, GraphProcedure procedure,
-                         Traversal.Parameters params, Set<Retrievable> filter) {
+                         GraphTraversal.Parameters params, Set<Retrievable> filter) {
         assert procedure.edgesCount() > 0;
         this.graphMgr = graphMgr;
         this.procedure = procedure;
