+++ traversal/procedure/ProcedureVertex.java
@@ -28,7 +28,7 @@ import com.vaticle.typedb.core.graph.vertex.AttributeVertex;
 import com.vaticle.typedb.core.graph.vertex.ThingVertex;
 import com.vaticle.typedb.core.graph.vertex.TypeVertex;
 import com.vaticle.typedb.core.graph.vertex.Vertex;
-import com.vaticle.typedb.core.traversal.Traversal;
+import com.vaticle.typedb.core.traversal.GraphTraversal;
 import com.vaticle.typedb.core.traversal.common.Identifier;
 import com.vaticle.typedb.core.traversal.graph.TraversalVertex;
 import com.vaticle.typedb.core.traversal.predicate.Predicate;
@@ -72,7 +72,7 @@ public abstract class ProcedureVertex<
         this.dependedEdgeOrders = new AtomicReference<>(null);
     }
 
-    public abstract FunctionalIterator<? extends VERTEX> iterator(GraphManager graphMgr, Traversal.Parameters parameters);
+    public abstract FunctionalIterator<? extends VERTEX> iterator(GraphManager graphMgr, GraphTraversal.Parameters parameters);
 
     @Override
     public void in(ProcedureEdge<?, ?> edge) {
@@ -140,7 +140,7 @@ public abstract class ProcedureVertex<
         public ProcedureVertex.Thing asThing() { return this; }
 
         @Override
-        public FunctionalIterator<? extends ThingVertex> iterator(GraphManager graphMgr, Traversal.Parameters parameters) {
+        public FunctionalIterator<? extends ThingVertex> iterator(GraphManager graphMgr, GraphTraversal.Parameters parameters) {
             assert isStartingVertex();
             if (props().hasIID()) return iterateAndFilterFromIID(graphMgr, parameters);
             else if (!props().types().isEmpty()) return iterateAndFilterFromTypes(graphMgr, parameters);
@@ -152,7 +152,7 @@ public abstract class ProcedureVertex<
         }
 
         FunctionalIterator<? extends ThingVertex> filter(FunctionalIterator<? extends ThingVertex> iterator,
-                                                         Traversal.Parameters params) {
+                                                         GraphTraversal.Parameters params) {
             if (props().hasIID()) iterator = filterIID(iterator, params);
             if (!props().types().isEmpty()) iterator = filterTypes(iterator);
             if (!props().predicates().isEmpty()) iterator = filterPredicates(filterAttributes(iterator), params);
@@ -176,7 +176,7 @@ public abstract class ProcedureVertex<
         }
 
         private FunctionalIterator<? extends AttributeVertex<?>> iterateAndFilterFromAttributes(
-                GraphManager graph, Traversal.Parameters parameters) {
+                GraphManager graph, GraphTraversal.Parameters parameters) {
             FunctionalIterator<? extends AttributeVertex<?>> iter;
             FunctionalIterator<TypeVertex> attTypes;
 
@@ -205,7 +205,7 @@ public abstract class ProcedureVertex<
             return tree(rootType, t -> t.ins().edge(SUB).from()).flatMap(t -> graphMgr.data().getReadable(t));
         }
 
-        FunctionalIterator<? extends ThingVertex> iterateAndFilterFromIID(GraphManager graphMgr, Traversal.Parameters parameters) {
+        FunctionalIterator<? extends ThingVertex> iterateAndFilterFromIID(GraphManager graphMgr, GraphTraversal.Parameters parameters) {
             assert props().hasIID() && id().isVariable();
             Identifier.Variable id = id().asVariable();
             FunctionalIterator<? extends ThingVertex> iter = single(graphMgr.data().getReadable(parameters.getIID(id))).noNulls();
@@ -215,7 +215,7 @@ public abstract class ProcedureVertex<
         }
 
         FunctionalIterator<? extends ThingVertex> iterateAndFilterFromTypes(GraphManager graphMgr,
-                                                                            Traversal.Parameters parameters) {
+                                                                            GraphTraversal.Parameters parameters) {
             assert !props().types().isEmpty();
             FunctionalIterator<? extends ThingVertex> iter;
             Optional<Predicate.Value<?>> eq = iterate(props().predicates()).filter(p -> p.operator().equals(EQ)).first();
@@ -235,12 +235,12 @@ public abstract class ProcedureVertex<
         }
 
         FunctionalIterator<? extends ThingVertex> filterIID(FunctionalIterator<? extends ThingVertex> iterator,
-                                                            Traversal.Parameters parameters) {
+                                                            GraphTraversal.Parameters parameters) {
             return iterator.filter(v -> v.iid().equals(parameters.getIID(id().asVariable())));
         }
 
         FunctionalIterator<ThingEdge> filterIIDOnEdge(FunctionalIterator<ThingEdge> iterator,
-                                                      Traversal.Parameters parameters, boolean isForward) {
+                                                      GraphTraversal.Parameters parameters, boolean isForward) {
             Function<ThingEdge, ThingVertex> fn = e -> isForward ? e.to() : e.from();
             return iterator.filter(e -> fn.apply(e).iid().equals(parameters.getIID(id().asVariable())));
         }
@@ -255,19 +255,19 @@ public abstract class ProcedureVertex<
         }
 
         FunctionalIterator<? extends AttributeVertex<?>> filterPredicates(FunctionalIterator<? extends AttributeVertex<?>> iterator,
-                                                                          Traversal.Parameters parameters) {
+                                                                          GraphTraversal.Parameters parameters) {
             return filterPredicates(iterator, parameters, null);
         }
 
         FunctionalIterator<? extends AttributeVertex<?>> filterPredicates(FunctionalIterator<? extends AttributeVertex<?>> iterator,
-                                                                          Traversal.Parameters parameters,
+                                                                          GraphTraversal.Parameters parameters,
                                                                           @Nullable Predicate.Value<?> exclude) {
             // TODO: should we throw an exception if the user asserts a value predicate on a non-attribute?
             // TODO: should we throw an exception if the user assert a value non-comparable value types?
             assert id().isVariable();
             for (Predicate.Value<?> predicate : props().predicates()) {
                 if (Objects.equals(predicate, exclude)) continue;
-                for (Traversal.Parameters.Value value : parameters.getValues(id().asVariable(), predicate)) {
+                for (GraphTraversal.Parameters.Value value : parameters.getValues(id().asVariable(), predicate)) {
                     iterator = iterator.filter(a -> predicate.apply(a, value));
                 }
             }
@@ -275,12 +275,12 @@ public abstract class ProcedureVertex<
         }
 
         FunctionalIterator<ThingEdge> filterPredicatesOnEdge(FunctionalIterator<ThingEdge> iterator,
-                                                             Traversal.Parameters parameters, boolean isForward) {
+                                                             GraphTraversal.Parameters parameters, boolean isForward) {
             assert id().isVariable();
             Function<ThingEdge, ThingVertex> fn = e -> isForward ? e.to() : e.from();
             iterator = iterator.filter(e -> fn.apply(e).isAttribute());
             for (Predicate.Value<?> predicate : props().predicates()) {
-                for (Traversal.Parameters.Value value : parameters.getValues(id().asVariable(), predicate)) {
+                for (GraphTraversal.Parameters.Value value : parameters.getValues(id().asVariable(), predicate)) {
                     iterator = iterator.filter(e -> predicate.apply(fn.apply(e).asAttribute(), value));
                 }
             }
@@ -288,7 +288,7 @@ public abstract class ProcedureVertex<
         }
 
         FunctionalIterator<? extends AttributeVertex<?>> iteratorOfAttributesWithTypes(
-                GraphManager graphMgr, Traversal.Parameters params, Predicate.Value<?> eq) {
+                GraphManager graphMgr, GraphTraversal.Parameters params, Predicate.Value<?> eq) {
             FunctionalIterator<TypeVertex> attributeTypes = iterate(props().types().iterator())
                     .map(l -> graphMgr.schema().getType(l)).noNulls()
                     .map(t -> {
@@ -300,16 +300,16 @@ public abstract class ProcedureVertex<
 
         FunctionalIterator<? extends AttributeVertex<?>> iteratorOfAttributes(
                 GraphManager graphMgr, FunctionalIterator<TypeVertex> attributeTypes,
-                Traversal.Parameters parameters, Predicate.Value<?> eqPredicate) {
+                GraphTraversal.Parameters parameters, Predicate.Value<?> eqPredicate) {
             // TODO: should we throw an exception if the user asserts 2 values for a given vertex?
             assert id().isVariable();
-            Set<Traversal.Parameters.Value> values = parameters.getValues(id().asVariable(), eqPredicate);
+            Set<GraphTraversal.Parameters.Value> values = parameters.getValues(id().asVariable(), eqPredicate);
             assert values.size() == 1;
             return attributeTypes.map(t -> attributeVertex(graphMgr, t, values.iterator().next())).noNulls();
         }
 
         private AttributeVertex<?> attributeVertex(GraphManager graphMgr, TypeVertex type,
-                                                   Traversal.Parameters.Value value) {
+                                                   GraphTraversal.Parameters.Value value) {
             assert type.isAttributeType();
             switch (type.valueType()) {
                 case BOOLEAN:
@@ -344,7 +344,7 @@ public abstract class ProcedureVertex<
         }
 
         @Override
-        public FunctionalIterator<TypeVertex> iterator(GraphManager graphMgr, Traversal.Parameters parameters) {
+        public FunctionalIterator<TypeVertex> iterator(GraphManager graphMgr, GraphTraversal.Parameters parameters) {
             assert isStartingVertex() && id().isVariable();
             FunctionalIterator<TypeVertex> iterator = null;
